requirements.txt
Flask==3.0.3
imutils==0.5.4
Keras==3.3.3
opencv-python==4.10.0
Pillow==8.0.1
tensorflow==2.17.0
scikit-image==0.24.0
